requirements.txt
langchain
flask
flask-cors
streamlit
pytesseract
deep-translator
requests
Pillow
uvicorn
